chore: promote create --ensure-stack-ids
